change: move HTTP_BIN_URL to utils
